adding remote storage delete backup after transfer option

+++ includes/class-xcloner-scheduler.php
@@ -325,7 +325,7 @@ class Xcloner_Scheduler{
 				$this->archive_system->send_notification($to, $from, $subject, $return['extra']['backup_parent'], $schedule, "", $additional);
 				
 				//CHECK IF WE SHOULD DELETE BACKUP AFTER REMOTE TRANSFER IS DONE
-				if($this->xcloner_settings->get_xcloner_option('xcloner_cleanup_delete_after_remote_transfer'))
+				if($schedule['remote_storage'] && $this->xcloner_settings->get_xcloner_option('xcloner_cleanup_delete_after_remote_transfer'))
 				{
 					$this->logger->info(sprintf("Deleting %s from local storage matching rule xcloner_cleanup_delete_after_remote_transfer",$return['extra']['backup_parent']));
 					$this->xcloner_file_system->delete_backup_by_name($return['extra']['backup_parent']);
